Implement image upload. Organize html Pages
Implement image upload on server.Image upload folder moved
from App_Data to root. Add appropriate imageUrl
in js. Profile now contains orders,reviews,products in a single
page. Implement product ordering and leaving a review on order complete.

+++ Projekat/Repository/Impl/ProductRepository.cs
@@ -26,16 +26,16 @@ namespace Projekat.Repository
         {
             message = string.Empty;
             User user = userDao.FindById(userId);
-            if(user == default(User))
+            if (user == default(User))
             {
                 message = "User not found!";
                 return Enumerable.Empty<Product>();
             }
-            if(user.Role == UserType.Buyer)
+            if (user.Role == UserType.Buyer)
             {
                 return productDao.FindByIds(user.Favourites);
             }
-            else if(user.Role == UserType.Seller)
+            else if (user.Role == UserType.Seller)
             {
                 return productDao.FindByIds(user.PublishedProducts);
             }
@@ -45,7 +45,7 @@ namespace Projekat.Repository
         {
             message = string.Empty;
             User seller = userDao.FindById(product.SellerId);
-            if(seller  == default(User))
+            if (seller == default(User))
             {
                 message = "Seller not found";
                 return product;
@@ -57,16 +57,74 @@ namespace Projekat.Repository
             return added;
         }
 
+        public IEnumerable<int> AddProductToFavourites(int userId, int productId, out string message)
+        {
+            message = string.Empty;
+            User user = userDao.FindById(userId);
+            if (user == default(User))
+            {
+                message = "User not found!";
+                return Enumerable.Empty<int>();
+            }
+            Product product = productDao.FindById(productId);
+            if (product == default(Product))
+            {
+                message = "Product not found";
+                return Enumerable.Empty<int>();
+            }
+            if (user.Favourites.Contains(productId))
+            {
+                user.Favourites.Remove(productId);
+            }
+            else
+            {
+                user.Favourites.Add(productId);
+            }
+            return user.Favourites;
+        }
+
         public Product DeleteProduct(int productId)
         {
+            Product product = productDao.FindById(productId);
+            if(product == default(Product))
+            {
+                return product;
+            }
+            User user = userDao.FindById(product.SellerId);
+            user.PublishedProducts.Remove(productId);
+            userDao.RemoveProductFromFav(productId);
             return productDao.DeleteProduct(productId);
         }
 
         public Product UpdateProduct(Product updatedProduct, out string message)
         {
             message = string.Empty;
-            // TODO: Possible validation checks
-            return productDao.UpdateProduct(updatedProduct);
+            if(productDao.FindById(updatedProduct.ID) == default(Product))
+            {
+                message = "Product not found";
+                return default(Product);
+            }
+            User seller = userDao.FindById(updatedProduct.SellerId);
+            if (seller == default(User))
+            {
+                message = "Seller not found";
+                return updatedProduct;
+            }
+            if(updatedProduct.Price <= 0)
+            {
+                message = "Price must be greater than 0";
+                return default(Product);
+            }
+            if(updatedProduct.Amount < 0)
+            {
+                message = "Amount can not be negative";
+                return default(Product);
+            }
+            // TODO: Check City against a list of available cities
+            updatedProduct.PublishDate = DateTime.UtcNow;
+            Product updated = productDao.UpdateProduct(updatedProduct);
+
+            return updated;
         }
     }
 }
